Restructured file architecture, tests and separated components into separate files
