+++ app/src/main/java/com/example/therapet/app/auth/ResetPasswordScreen.kt
@@ -18,10 +18,17 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.therapet.app.ui.components.BasicTopBar
-import com.example.therapet.app.ui.components.MyFilledButton
+import com.example.therapet.app.ui.components.bars.BasicTopBar
 import com.example.therapet.app.ui.theme.TheraPetTheme
 import com.example.therapet.R
+import com.example.therapet.app.ui.components.buttons.general.CustomFilledButton
+
+/**
+ * @author: Chloe Edwards
+ * @date: 24/12/2025
+ *
+ * Reset password screen UI
+ */
 
 @Composable
 fun ResetPasswordScreen(
@@ -88,7 +95,7 @@ fun ResetPasswordScreen(
 
 @Composable
 fun ResetPasswordButton(onClick: () -> Unit){
-    MyFilledButton(
+    CustomFilledButton(
         onClick = onClick,
         text = stringResource(R.string.reset_password_title),
         modifier = Modifier
