+++ app/src/main/java/com/example/therapet/app/auth/WelcomeScreen.kt
@@ -17,7 +17,14 @@ import androidx.compose.ui.unit.dp
 import com.example.therapet.app.ui.theme.TheraPetTheme
 import androidx.compose.ui.res.stringResource
 import com.example.therapet.R
-import com.example.therapet.app.ui.components.MyFilledButton
+import com.example.therapet.app.ui.components.buttons.general.CustomFilledButton
+
+/**
+ * @author: Chloe Edwards
+ * @date: 24/12/2025
+ *
+ * Welcome screen UI
+ */
 
 //Composing the full screen
 @Composable
@@ -59,7 +66,7 @@ fun WelcomeScreen(
 //Register button
 @Composable
 fun ChooseRegisterButton(onClick: () -> Unit){
-    MyFilledButton(
+    CustomFilledButton(
         onClick = onClick,
         text = stringResource(R.string.register),
         modifier = Modifier
@@ -72,7 +79,7 @@ fun ChooseRegisterButton(onClick: () -> Unit){
 //Login button
 @Composable
 fun ChooseLoginButton(onClick: () -> Unit){
-    MyFilledButton(
+    CustomFilledButton(
         onClick = onClick,
         text = stringResource(R.string.login),
         modifier = Modifier
