+++ app/src/main/java/com/example/therapet/app/nav/Routes.kt
@@ -1,0 +1,27 @@
+package com.example.therapet.app.nav
+
+/**
+ * @Author: Chloe Edwards
+ * @Date: 21/12/2025
+ *
+ * Compiling routes into a single object - used by NavGraph
+ */
+
+object Routes {
+    const val WELCOME = "welcome_screen"
+    const val REGISTER = "register"
+    const val LOGIN = "login"
+    const val CREATE_PET = "create_pet"
+    const val HOME = "home_screen"
+    const val PROFILE = "profile_screen"
+    const val SETTINGS = "settings_screen"
+    const val APPOINTMENTS = "appointments_screen"
+    const val CHOOSE_THERAPIST = "choose_therapist_screen"
+    const val BOOK_APPOINTMENT = "book_appointment_screen"
+    const val RESET_PASSWORD = "reset_password_screen"
+    const val PET_SETTINGS = "pet_settings_screen"
+    const val PRIVACY_POLICY = "privacy_policy_screen"
+    const val HELP_SUPPORT = "help_support_screen"
+    const val DELETE_ACCOUNT = "delete_account_screen"
+    const val DELETE_ACCOUNT_CONFIRM = "delete_account_confirm_screen"
+}
