+++ app/src/main/java/com/example/therapet/app/ui/components/bars/BasicTopBar.kt
@@ -1,0 +1,86 @@
+package com.example.therapet.app.ui.components.bars
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.therapet.app.ui.theme.TheraPetTheme
+
+/**
+ * @Author: Chloe Edwards
+ * @Date: 24/12/2025
+ *
+ * A basic top bar with a nullable back bar, reusable on any screen
+ */
+
+@OptIn(ExperimentalMaterial3Api::class)
+
+@Composable
+fun BasicTopBar(
+    text: String,
+    onBackClick: (() -> Unit)? = null,
+) {
+    CenterAlignedTopAppBar(
+        colors = TopAppBarDefaults.topAppBarColors(
+            containerColor = MaterialTheme.colorScheme.primaryContainer,
+            titleContentColor = MaterialTheme.colorScheme.primary,
+        ),
+        title = {
+            Text(
+                text = text,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                style = MaterialTheme.typography.titleMedium
+            )
+        },
+        navigationIcon = {
+            if(onBackClick != null) {
+                IconButton(
+                    onClick = onBackClick,
+                    modifier = Modifier.testTag("back_button")
+                ) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                        contentDescription = "Back",
+                    )
+                }
+            }
+        },
+        actions = { /* There are no icons on this bar so no actions are needed */ }
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun BasicTopBarPreview() {
+    TheraPetTheme {
+
+        Column {
+            BasicTopBar(
+                text = "With Back Button",
+                onBackClick = {},
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            BasicTopBar(
+                text = "Without Back Button",
+                onBackClick = null,
+            )
+        }
+    }
+}
