+++ app/src/main/java/com/example/therapet/app/ui/components/buttons/general/CustomTonalFilledButton.kt
@@ -1,0 +1,49 @@
+package com.example.therapet.app.ui.components.buttons.general
+
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.FilledTonalButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
+import com.example.therapet.app.ui.theme.TheraPetTheme
+
+/**
+ * @Author: Chloe Edwards
+ * @Date: 23/12/2025
+ *
+ * A custom filled tonal button using the FilledTonal button object
+ */
+
+@Composable
+fun CustomTonalFilledButton(
+    modifier: Modifier = Modifier,
+    text: String,
+    onClick: () -> Unit,
+    containerColor: Color = MaterialTheme.colorScheme.primary,
+    contentColor: Color = MaterialTheme.colorScheme.onPrimary
+){
+    FilledTonalButton(
+        onClick = onClick,
+        modifier = modifier,
+        colors = ButtonDefaults.filledTonalButtonColors(
+            containerColor = containerColor,
+            contentColor = contentColor
+        )
+    ) {
+        Text(
+            text,
+            style = MaterialTheme.typography.labelMedium
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun CustomTonalFilledButtonPreview() {
+    TheraPetTheme {
+        CustomTonalFilledButton(onClick = {}, text = "Tonal Filled")
+    }
+}
