+++ app/src/main/java/com/example/therapet/app/ui/components/fields/input/OutlinedTextField.kt
@@ -1,0 +1,54 @@
+package com.example.therapet.app.ui.components.fields.input
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.therapet.app.ui.theme.TheraPetTheme
+
+/**
+ * @Author: Chloe Edwards
+ * @Date: 24/12/2025
+ *
+ * A custom outlined text field using OutlinedTextField
+ */
+
+@Composable
+fun CustomOutlinedTextField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholder: String,
+    label: String,
+    modifier: Modifier = Modifier,
+    textStyle: TextStyle = MaterialTheme.typography.labelSmall
+) {
+    OutlinedTextField(
+        value = value,
+        onValueChange = onValueChange,
+        modifier = modifier
+            .height(60.dp)
+            .fillMaxWidth(),
+        placeholder = { Text(placeholder, style = textStyle) },
+        label = { Text(label, style = textStyle) },
+        textStyle = textStyle,
+        singleLine = true
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun CustomOutlinedTextFieldPreview() {
+    TheraPetTheme {
+        CustomOutlinedTextField(
+            value = "Outlined Text Field",
+            onValueChange = {},
+            placeholder = "Outlined Text field",
+            label = " Outlined Text field label")
+    }
+}
